requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
Django==4.2.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
pycodestyle==2.10.0
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3